Update global.ts
As @Gusarich mentioned the fact that we will get a syntax error on passing invalid string to `toNano` function I decided to make this change to let the input contain separators like comma and underscore. So there would be no error using those separators to make your numbers more readable.