feat/fix/refactor: full rewrite of constant evaluator (#432)
to fix various issues and support a larger Tact expression fragment.
It is implemented as a classical tree-traversing interpreter.
We take into account the semantics of TVM operations, including
integer overflows.
As an optimizer it is not very efficient, though. But this will
be handled separately.
Supported conditional expressions, unboxing operator,
struct instances and struct field access.
Slices are not handled.