refactor(parseImports): new JustImports parser
It tries to parse the imports at the start of the source file
and then kind of ignores the rest of the input.
I don't have the Ohm.js docs on the plane, so I cannot check
if there is a special function to ignore the rest of the input
more efficiently.
Basically, it would be nice to have a parser that does not
have to consume the whole input returning the successfully
parsed part and the rest of the input, just like parser
combinators usually do.