Fix Typo in CHANGELOG.md
In Roadmap.md there was a line "let". The mistake is that the first letter in the sentence should be capitalized. 

I suggest fixing this problem by writing "Let" instead of "let".

Original:
"let-statements can now be used without an explicit type declaration and determine the type automatically if it was not specified: PR #198 and PR #438"
Improved:
"Let-statements can now be used without an explicit type declaration and determine the type automatically if it was not specified: PR #198 and PR #438"

Closes #591

- [ ] I have updated CHANGELOG.md
- [ ] I have documented my contribution in Tact Docs: https://github.com/tact-lang/tact-docs/pull/PR-NUMBER
- [ ] I have added tests to demonstrate the contribution is correctly implemented: this usually includes both positive and negative tests, showing the happy path(s) and featuring intentionally broken cases
- [ ] I have run all the tests locally and no test failure was reported
- [ ] I have run the linter, formatter and spellchecker
- [ ] I did not do unrelated and/or undiscussed refactorings