feat: enhance `asm` functions to support all of the Fift-asm syntax
Note, that while it's possible to support the syntax via Ohm without
constucting ASTs, it's impossible to support the AST generation without
outsourcing that part of the parser to the recursive-descent one, which
would also keep track of dictionary of words/instructions and the TVM
stack, at least approximating it. [See this for more
info](https://github.com/tact-lang/tact/issues/837#issuecomment-2362498612).