test: add Fift libraries from TON Blockchain monorepo
And apply a small fix to the parser — we cannot prohibit shadowing or
removal of instructions because both are used in Fift library and
beyond.

However, that means that out of place `{` and `}` are possible when
shadowing them. See the comment in `embed-fift-fif.tact` about such
case, and why it's ok to move forward with it