feat: enhance `asm` functions to support all* of the Fift-asm syntax (#855)
Note, that while it's possible to support the syntax via Ohm without
constucting ASTs, it's impossible to support the AST generation without
outsourcing that part of the parser to the recursive-descent one, which
would also keep track of dictionary of words/instructions and the TVM
stack, at least approximating it. [See this for more
info](https://github.com/tact-lang/tact/issues/837#issuecomment-2362498612).

* test: add Fift libraries from TON Blockchain monorepo

And apply a small fix to the parser — we cannot prohibit shadowing or
removal of instructions because both are used in Fift library and
beyond.

However, that means that out of place `{` and `}` are possible when
shadowing them. See the comment in `embed-fift-fif.tact` about such
case, and why it's ok to move forward with it