doc: migration of `tact-docs` to this repo and to Starlight framework (#880)
* feat: adjusted styles, added fonts and styles for KaTeX

* feat: configuration of docs

* feat: grammars for code highlighting and an impromptu plugin for inline code highlighting

A single-file rehype plugin, which adds inline code highlighting. Works
best for dark theme now, will be improved in the future for better
support of light theme

* feat(CI): testing and deployment workflows

The deployment one can be slightly adjusted to allow deployments on each
merge to master. For actual deployments, we've got to setup GitHub Pages
here and disable them in `tact-docs` repo