fix: bit shift FunC compilation errors for incorrect bit widths
This will only work for the simple cases as in the audit report, i.e.
when the bit widths are immediate (in-place) constant expressions.
We are going to implement a proper constant propagation static
analysis in the next Tact release.

Added extra tests with constant and function evaluation

Initial commit: 32dbaa8d84d200903c2c3321cc168707b7b832b7

Fixes #1383